server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
openai>=1.57.0
lunar-python==1.3.9
httpx>=0.28.0
